Extract `AnimationFrameCallback` from NativeProxy.h
