+++ packages/react-native-reanimated/android/src/main/cpp/reanimated/android/NativeProxy.h
@@ -21,31 +21,6 @@ namespace reanimated {
 using namespace facebook;
 using namespace facebook::jni;
 
-class AnimationFrameCallback : public HybridClass<AnimationFrameCallback> {
- public:
-  static auto constexpr kJavaDescriptor =
-      "Lcom/swmansion/reanimated/nativeProxy/AnimationFrameCallback;";
-
-  void onAnimationFrame(double timestampMs) {
-    callback_(timestampMs);
-  }
-
-  static void registerNatives() {
-    javaClassStatic()->registerNatives({
-        makeNativeMethod(
-            "onAnimationFrame", AnimationFrameCallback::onAnimationFrame),
-    });
-  }
-
- private:
-  friend HybridBase;
-
-  explicit AnimationFrameCallback(std::function<void(double)> callback)
-      : callback_(std::move(callback)) {}
-
-  std::function<void(double)> callback_;
-};
-
 class EventHandler : public HybridClass<EventHandler> {
  public:
   static auto constexpr kJavaDescriptor =
