+++ packages/react-native-reanimated/android/src/main/cpp/reanimated/android/OnLoad.cpp
@@ -1,5 +1,6 @@
 #include <fbjni/fbjni.h>
 
+#include <reanimated/android/AnimationFrameCallback.h>
 #include <reanimated/android/NativeProxy.h>
 
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM *vm, void *) {
